Add IRC color formatting to the alternate compact buffer list
IRC color formatting lets clients add simple formatting to their
messages and is widely implemented.

Message formatting is not appropriate for the modern look & feel of the
default message list, which looks like those of other modern IM systems
where this kind of formatting is not available. With colored bubbles as
the background of message text, colored messages would not look great by
default.

Now that there is an alternate compact message view that is disabled by
default, users that want to have a more traditional IRC experience can
see messages in a look and feel that is close to usual IRC clients.

In that context, adding IRC message formatting support to that alternate
message view enables power users to see basic markup and colors just
like they do on their other clients. We do not use coloured bubbles
either so any formatting will look nicer than on the normal list.

We intentionally support only a basic subset of the available message
formatting in order to keep that feature simple and small, as most
of the advanced formatting features are not widely used.

See: https://modern.ircdocs.horse/formatting.html