Fix buffers and notifications opened with no reason
When synchronizing in the background, we ask Android to keep our
process alive until we're doing fetching chat history from the
server. To do so, we only switch NetworkModel.state to online
when the end of the chat history batch is received.

However, this isn't quite enough: we still need to perform some
processing after the end of the chat history batch is received.
We need to store messages in the DB and perform other book-keeping.
This creates a race: our book-keeping vs. Android killing the
network connection.

As a result the connection to the IRC server may be killed before
we finish our processing. This resets CHANTYPES and makes
Client.isChannel always return false. ClientController then
mistakenly interprets all messages as direct messages.

Fix this by waiting for _handleChatMessages to complete before
switching NetworkModel.state to online.

Closes: https://todo.sr.ht/~emersion/goguma/82