Add a network_details page
This page is intended to be the network counterpart of the
buffer_details page. It enables a user to review the main information
about a network (only name and state for now, perhaps some attributes in
the future).

The list of conversations opened with this network is also listed, with
links to open the corresponding conversation when tapped. Conversely,
the existing mentions of networks (in settings and buffer_details) were
linked to open this page.

This page also enables the user to run two actions:
- editing the network (this opens the edit_network page)
- deleting the network (this simply sends a DELNETWORK and pops the
  Navigator)