Order buffers and buffer members in a case-insensitive order
Previously, we would list nicks with an uppercase initial letter before
those with a lowercase nick.

This changes the logic so that we sort nicks in a case-insensitive
manner. There are no case insensitive compare functions in the Flutter
stdlib so we have to do a toLowerCase comparison. This is only run once
when opening buffer details so it should be OK wrt performance.

Co-authored-by: Simon Ser <contact@emersion.fr>