push: fix tapping notification when app is running
Before this commit, tapping a notification while the app is running
would not open the target buffer.

This is due to issues with Isolates [1]. The push notifications are
handled in a separate Isolate, and initializing the notification
plugin in that Isolate makes it so onDidReceiveNotificationResponse
is invoked in the push Isolate instead of the main Isolate.

Fix this by working around the issue and avoid initializing the
notification plugin in the push Isolate until a better fix is merged
upstream [2].

[1]: https://github.com/MaikuB/flutter_local_notifications/issues/1730
[2]: https://github.com/MaikuB/flutter_local_notifications/pull/1744