Introduce WebPushSubscriptionEntry.tag
This allows push providers to attach an arbitrary tag to the
subscription. Firebase uses this to store the pushgarden
subscription URL, UnifiedPush uses this to store the instance.

This allows for more robust push notification handling. We no
longer get confused when a stale push notification comes in (because
it's delayed, or comes from a previous installation).