Fix regression in compact message formatting
A recent commit [1] broke mesage formatting in compact message mode.

Linkify was stripping all formatting, and the style was never passed
back before displaying it.

We can just wrap the TextSpan linkify returns in another TextSpan with
the current span style and with an only child that is the linkified
text span.

[1]: 6c7c51b029c9be8664c958279f632313bac5f193