Do not crop images horizontally
While making the image height fixed is important to avoid layout
changes, there is no impact to letting the image take as much
width as it needs, dynamically.

For images that are wider than they are tall, this enables the user
to see the full image without cropping the left and right parts of
the image.

Tested locally on both compact and non-compact modes.