page/buffer: Make invisible compact message spacing unselectable
We're building compact message items by writing an invisible prefix
(stacked with an actual username), then the actual message, then
an invisible suffix (stack with an actual timestamp).

Currenty selecting a compact message text results in selecting the
invisible prefixes and suffixes as well, which results in unexpected
data being selected.

This patch makes these invisible hacky spacings unselectable so that
only the actual message can be selected and copied in the list.